✨ now canvas does not keep texture data and size, texture object with it's data and properties should be passed by the owner

+++ src/lib/LCD/Simulator/LCDRenderer.h
@@ -10,8 +10,8 @@ typedef struct {
 
 void LCDSetDisplayArea(UINT16 Xstart, UINT16 Ystart, UINT16 Xend, UINT16 Yend);
 void LCDClear(UINT16 FillColor);
-void LCDRenderTexture(UINT16 *texture);
-void LCDRenderTextureInArea(UINT16 xStart, UINT16 yStart, UINT16 xEnd, UINT16 yEnd, UINT16 *texture);
+void LCDRenderTexture(UINT8 *texture);
+void LCDRenderTextureInArea(UINT16 xStart, UINT16 yStart, UINT16 xEnd, UINT16 yEnd, UINT8 *texture);
 void LCDRenderPoint(UINT16 x, UINT16 y, UINT16 color);
 void LCDRenderPng(FIL *file);
 bool LCDRenderShouldClose();
